@@ -2,10 +2,10 @@ package br.com.alura.forum.config.security;
 
 import java.io.IOException;
 
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -31,12 +31,10 @@ public class AutenticacaoViaTokenFilter extends OncePerRequestFilter {
 		String token = recuperarToken(request);
 		boolean valido = tokenService.isTokenValido(token);
 		if (valido) {
-			autenticarCliente(token); 
-				
+			autenticarCliente(token);
 		}
 		
 		filterChain.doFilter(request, response);
-		
 	}
 
 	private void autenticarCliente(String token) {
