@@ -4,18 +4,19 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.web.SecurityFilterChain;
 
 @EnableWebSecurity
 @Configuration
 @Profile("dev")
-public class DevSecurityConfigurations extends WebSecurityConfigurerAdapter {
+public class DevSecurityConfigurations {
 
-	@Override
-	protected void configure(HttpSecurity http) throws Exception {
-		http.authorizeRequests()
-		.antMatchers("/**").permitAll()
+	public SecurityFilterChain configure(HttpSecurity http) throws Exception {
+		http.authorizeHttpRequests()
+		.requestMatchers("/**").permitAll()
 		.and().csrf().disable();
+
+		return http.build();
 	}
 	
 }
