@@ -1,18 +1,18 @@
 package br.com.alura.forum;
 
-import org.junit.Assert;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.context.ActiveProfiles;
 
-@RunWith(SpringRunner.class)
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
 @SpringBootTest
+@ActiveProfiles("test")
 public class ForumApplicationTests {
 
 	@Test
 	public void contextLoads() {
-		Assert.assertTrue(true);
+		assertTrue(true);
 	}
 
 }
