@@ -1,21 +1,18 @@
 package br.com.alura.forum.controller;
 
-import java.net.URI;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 
-@RunWith(SpringRunner.class)
-@SpringBootTest
+import java.net.URI;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @AutoConfigureMockMvc
 @ActiveProfiles("test")
 public class AutenticacaoControllerTest {
@@ -28,15 +25,13 @@ public class AutenticacaoControllerTest {
 		URI uri = new URI("/auth");
 		String json = "{\"email\":\"invalido@email.com\",\"senha\":\"123456\"}";
 		
-		mockMvc
-		.perform(MockMvcRequestBuilders
-				.post(uri)
-		        .content(json)
-		        .contentType(MediaType.APPLICATION_JSON))
-		.andExpect(MockMvcResultMatchers
-				.status()
-				.is(400));
-		
+		mockMvc.perform(MockMvcRequestBuilders
+						.post(uri)
+						.content(json)
+						.contentType(MediaType.APPLICATION_JSON))
+				.andExpect(MockMvcResultMatchers
+						.status()
+						.is(400));
 	}
 
 }
