@@ -1,19 +1,16 @@
 package br.com.alura.forum.repository;
-import org.junit.Assert;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
 import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.junit4.SpringRunner;
 
 import br.com.alura.forum.modelo.Curso;
 
-@RunWith(SpringRunner.class)
+import static org.junit.jupiter.api.Assertions.*;
+
 @DataJpaTest
-@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
 @ActiveProfiles("test")
 public class CursoRepositoryTest {
 	
@@ -25,7 +22,7 @@ public class CursoRepositoryTest {
 
 	@Test
 	public void deveriaCarregarUmCursoAoBuscarPeloSeuNome() {
-		String nomeCurso = "HTML 5";
+		String nomeCurso = "HTML 4";
 		
 		Curso html5 = new Curso();
 		html5.setNome(nomeCurso);
@@ -33,8 +30,8 @@ public class CursoRepositoryTest {
 		em.persist(html5);
 		
 		Curso curso = repository.findByNome(nomeCurso);
-		Assert.assertNotNull(curso);
-		Assert.assertEquals(nomeCurso, curso.getNome());
+		assertNotNull(curso);
+		assertEquals(nomeCurso, curso.getNome());
 		
 	}
 	
@@ -42,7 +39,7 @@ public class CursoRepositoryTest {
 	public void NaoDeveriaCarregarUmCursoCujoNomeNaoEstejaCadastrado() {
 		String nomeCurso = "JPA";
 		Curso curso = repository.findByNome(nomeCurso);
-		Assert.assertNull(curso);
+		assertNull(curso);
 	}
 
 }
